fix: 修正了不带参访问即可获得所有数据的bug

+++ src/main/java/com/ocelot/controller/CourseTableController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -26,7 +25,7 @@ public class CourseTableController {
     @Autowired
     CourseService courseService;
 
-//  查询课表
+    //  查询课表
     @RequestMapping(value = "/classtable", method = RequestMethod.POST)
     public JSONObject getClassTable(String studentId, String studentPassword, String schoolYear) throws IOException {
         //从studentLogin返回的JSON对象
@@ -38,49 +37,54 @@ public class CourseTableController {
         //从教务系统获取的课表
         JSONArray classArrayFromSystem;
 
-        //从数据库/Redis获取课表
-        List<Course> courseList = courseService.selectCourseTable(studentId, schoolYear);
-
-        /*判断有无课表,有则带状态码返回课表,无则进入系统获取*/
-        if (!courseList.isEmpty()) {//判空
-            String courseStr = JSON.toJSONString(courseList);
-            courseArray = JSONArray.parseArray(courseStr);
-            responseObject.put("data", courseArray);
-            responseObject.put("code", true);
-            return responseObject;
-        } else {
-//            判断学号密码是否正确输入
-            if(studentId.isEmpty()||studentPassword.isEmpty()||studentPassword.isBlank()||studentId.isBlank()){
-                responseObject.put("msg","请输入教务系统的账号密码!");
-                responseObject.put("code", false);
+        if (!studentId.isEmpty() || !studentId.isBlank()) {
+            //从数据库/Redis获取课表
+            List<Course> courseList = courseService.selectCourseTable(studentId, schoolYear);
+            /*判断有无课表,有则带状态码返回课表,无则进入系统获取*/
+            if (!courseList.isEmpty()) {//判空
+                String courseStr = JSON.toJSONString(courseList);
+                courseArray = JSONArray.parseArray(courseStr);
+                responseObject.put("data", courseArray);
+                responseObject.put("code", 200);
                 return responseObject;
-            }else{
-                //执行登陆
-                loginObject = SystemHandler.studentLogin(studentId, studentPassword);
-                //判断登陆状态
-                if (loginObject.get("code").equals(true)) {
-                    classArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
-                    //状态码永远在数组第0位
-                    String statusCode = classArrayFromSystem.getJSONObject(0).get("code").toString();
-                    //判断课表获取状态
-                    if (statusCode.equals("true")) {
-                        classArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
-                        courseService.addCourseTable(classArrayFromSystem, studentId);
-                        return getClassTable(studentId, schoolYear, studentPassword);
+            } else {
+//            判断学号密码是否正确输入
+                if (!studentPassword.isEmpty() || !studentPassword.isBlank()) {
+                    //执行登陆
+                    loginObject = SystemHandler.studentLogin(studentId, studentPassword);
+                    //判断登陆状态
+                    if (loginObject.get("code").equals(true)) {
+                        classArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
+                        //状态码永远在数组第0位
+                        String statusCode = classArrayFromSystem.getJSONObject(0).get("code").toString();
+                        //判断课表获取状态
+                        if (statusCode.equals("true")) {
+                            classArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
+                            courseService.addCourseTable(classArrayFromSystem, studentId);
+                            return getClassTable(studentId, schoolYear, studentPassword);
+                        } else {
+                            return classArrayFromSystem.getJSONObject(0);
+                        }
                     } else {
-                        return classArrayFromSystem.getJSONObject(0);
+                        logger.info("用户 [{}] 未登录", studentId);
+                        return loginObject;
                     }
                 } else {
-                    logger.info("用户 [{}] 未登录",studentId);
-                    return loginObject;
+                    responseObject.put("msg", "请输入教务系统的账号密码!");
+                    responseObject.put("code", 403);
+                    return responseObject;
                 }
             }
+        } else {
+            responseObject.put("msg", "请输入学号!");
+            responseObject.put("code", 403);
+            return responseObject;
         }
     }
 
-//    更新课表
+    //    更新课表
     @RequestMapping(value = "/classtable", method = RequestMethod.PUT)
-    public JSONObject updateClassTable(String studentId, String studentPassword, String schoolYear) throws IOException{
+    public JSONObject updateClassTable(String studentId, String studentPassword, String schoolYear) throws IOException {
         //从studentLogin返回的JSON对象
         JSONObject loginObject;
         //初始化函数返回的JSON对象
@@ -90,38 +94,50 @@ public class CourseTableController {
         //从教务系统获取的课表
         JSONArray classArrayFromSystem;
 
-        //执行登陆
-        loginObject = SystemHandler.studentLogin(studentId, studentPassword);
-        //判断登陆状态
-        if (loginObject.get("code").equals(true)) {
-            classArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
-            //状态码永远在数组第0位
-            String statusCode = classArrayFromSystem.getJSONObject(0).get("code").toString();
-            //判断课表获取状态
-            if (statusCode.equals("true")) {
-                classArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
-                responseObject = courseService.updateCourseTable(classArrayFromSystem, studentId, schoolYear);
-                return responseObject;
+        if (!studentId.isEmpty() || !studentId.isBlank() || !studentPassword.isEmpty() || !studentPassword.isBlank()) {
+            //执行登陆
+            loginObject = SystemHandler.studentLogin(studentId, studentPassword);
+            //判断登陆状态
+            if (loginObject.get("code").equals(true)) {
+                classArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
+                //状态码永远在数组第0位
+                String statusCode = classArrayFromSystem.getJSONObject(0).get("code").toString();
+                //判断课表获取状态
+                if (statusCode.equals("true")) {
+                    classArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
+                    responseObject = courseService.updateCourseTable(classArrayFromSystem, studentId, schoolYear);
+                    return responseObject;
+                } else {
+                    return classArrayFromSystem.getJSONObject(0);
+                }
             } else {
-                return classArrayFromSystem.getJSONObject(0);
+                logger.info("用户 [{}] 未登录", studentId);
+                return loginObject;
             }
-        } else {
-            logger.info("用户 [{}] 未登录",studentId);
-            return loginObject;
+        }else{
+            responseObject.put("msg", "请输入教务系统的账号密码!");
+            responseObject.put("code", 403);
+            return responseObject;
         }
     }
 
-//    删除课表
+    //    删除课表
     @RequestMapping(value = "/classtable", method = RequestMethod.DELETE)
-    public JSONObject deleteClassTable(@RequestBody JSONArray studentIdArray){
+    public JSONObject deleteClassTable(@RequestBody JSONArray studentIdArray) {
+        JSONObject responseObject = new JSONObject();
 //        构造一个存放学号的列表
-        List<Long> studentIdList = new ArrayList<>();
-        for(int i=0; i<studentIdArray.size(); i++){
+        if(!studentIdArray.isEmpty()){
+            List<Long> studentIdList = new ArrayList<>();
+            for (int i = 0; i < studentIdArray.size(); i++) {
 //            将学号格式化成Long,存入列表
-            String tmpStr = studentIdArray.getString(i);
-            studentIdList.add(Long.parseLong(tmpStr));
+                String tmpStr = studentIdArray.getString(i);
+                studentIdList.add(Long.parseLong(tmpStr));
+            }
+            responseObject = courseService.deleteCourseTable(studentIdList);
+        }else{
+            responseObject.put("msg", "请至少输入一个学号!");
+            responseObject.put("code", 403);
         }
-        JSONObject responseObject = courseService.deleteCourseTable(studentIdList);
         return responseObject;
     }
 }
