+++ src/main/java/com/ocelot/controller/ExaminationController.java
@@ -38,42 +38,48 @@ public class ExaminationController {
         //从教务系统获取的课表
         JSONArray examinationArrayFromSystem;
 
-        //从数据库/Redis获取课表
-        List<Examination> examinationList = examinationService.selectExamination(studentId, schoolYear);
+        if (!studentId.isEmpty() || !studentId.isBlank()) {
+            //从数据库/Redis获取成绩
+            List<Examination> examinationList = examinationService.selectExamination(studentId, schoolYear);
 
-        /*判断有无课表,有则带状态码返回课表,无则进入系统获取*/
-        if (!examinationList.isEmpty()) {//判空
-            String courseStr = JSON.toJSONString(examinationList);
-            examinationArray = JSONArray.parseArray(courseStr);
-            responseObject.put("data", examinationArray);
-            responseObject.put("code", true);
-            return responseObject;
-        } else {
-            if(studentId.isEmpty()||studentPassword.isEmpty()||studentPassword.isBlank()||studentId.isBlank()){
-                responseObject.put("msg","请输入教务系统的账号密码!");
-                responseObject.put("code", false);
+            /*判断有无成绩,有则带状态码返回成绩,无则进入系统获取*/
+            if (!examinationList.isEmpty()) {//判空
+                String courseStr = JSON.toJSONString(examinationList);
+                examinationArray = JSONArray.parseArray(courseStr);
+                responseObject.put("data", examinationArray);
+                responseObject.put("code", 200);
                 return responseObject;
-            }else {
-                //执行登陆
-                loginObject = SystemHandler.studentLogin(studentId, studentPassword);
-                //判断登陆状态
-                if (loginObject.get("code").equals(true)) {
-                    examinationArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
-                    //状态码永远在数组第0位
-                    String statusCode = examinationArrayFromSystem.getJSONObject(0).get("code").toString();
-                    //判断课表获取状态
-                    if (statusCode.equals("true")) {
-                        examinationArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
-                        examinationService.addExamination(examinationArrayFromSystem, studentId);
-                        return getExamination(studentId, schoolYear, studentPassword);
+            } else {
+                if(studentId.isEmpty()||studentPassword.isEmpty()||studentPassword.isBlank()||studentId.isBlank()){
+                    responseObject.put("msg","请输入教务系统的账号密码!");
+                    responseObject.put("code", 403);
+                    return responseObject;
+                }else {
+                    //执行登陆
+                    loginObject = SystemHandler.studentLogin(studentId, studentPassword);
+                    //判断登陆状态
+                    if (loginObject.get("code").equals(true)) {
+                        examinationArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
+                        //状态码永远在数组第0位
+                        String statusCode = examinationArrayFromSystem.getJSONObject(0).get("code").toString();
+                        //判断课表获取状态
+                        if (statusCode.equals("true")) {
+                            examinationArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
+                            examinationService.addExamination(examinationArrayFromSystem, studentId);
+                            return getExamination(studentId, schoolYear, studentPassword);
+                        } else {
+                            return examinationArrayFromSystem.getJSONObject(0);
+                        }
                     } else {
-                        return examinationArrayFromSystem.getJSONObject(0);
+                        logger.info("用户 [{}] 未登录", studentId);
+                        return loginObject;
                     }
-                } else {
-                    logger.info("用户 [{}] 未登录", studentId);
-                    return loginObject;
                 }
             }
+        }else{
+            responseObject.put("msg", "请输入学号!");
+            responseObject.put("code", 403);
+            return responseObject;
         }
     }
 
@@ -89,38 +95,52 @@ public class ExaminationController {
         //从教务系统获取的课表
         JSONArray examinationArrayFromSystem;
 
-        //执行登陆
-        loginObject = SystemHandler.studentLogin(studentId, studentPassword);
-        //判断登陆状态
-        if (loginObject.get("code").equals(true)) {
-            examinationArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
-            //状态码永远在数组第0位
-            String statusCode = examinationArrayFromSystem.getJSONObject(0).get("code").toString();
-            //判断课表获取状态
-            if (statusCode.equals("true")) {
-                examinationArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
-                responseObject = examinationService.updateExamination(examinationArrayFromSystem, studentId, schoolYear);
-                return responseObject;
+        if (!studentId.isEmpty() || !studentId.isBlank() || !studentPassword.isEmpty() || !studentPassword.isBlank()) {
+
+            //执行登陆
+            loginObject = SystemHandler.studentLogin(studentId, studentPassword);
+            //判断登陆状态
+            if (loginObject.get("code").equals(true)) {
+                examinationArrayFromSystem = SystemHandler.takeClassTable(Integer.parseInt(schoolYear));
+                //状态码永远在数组第0位
+                String statusCode = examinationArrayFromSystem.getJSONObject(0).get("code").toString();
+                //判断课表获取状态
+                if (statusCode.equals("true")) {
+                    examinationArrayFromSystem.remove(0);//判断为true后,将数组首位的状态码删除
+                    responseObject = examinationService.updateExamination(examinationArrayFromSystem, studentId, schoolYear);
+
+                } else {
+                    responseObject =  examinationArrayFromSystem.getJSONObject(0);
+                }
             } else {
-                return examinationArrayFromSystem.getJSONObject(0);
+                logger.info("用户 [{}] 未登录",studentId);
+                return loginObject;
             }
-        } else {
-            logger.info("用户 [{}] 未登录",studentId);
-            return loginObject;
+        }else{
+            responseObject.put("msg", "请输入教务系统的账号密码!");
+            responseObject.put("code", 403);
         }
+        return responseObject;
     }
 
     //    删除课表
     @RequestMapping(value = "/examination", method = RequestMethod.DELETE)
     public JSONObject deleteClassTable(@RequestBody JSONArray studentIdArray){
-//        构造一个存放学号的列表
-        List<Long> studentIdList = new ArrayList<>();
-        for(int i=0; i<studentIdArray.size(); i++){
+        JSONObject responseObject = new JSONObject();
+
+        if(!studentIdArray.isEmpty()){
+            //        构造一个存放学号的列表
+            List<Long> studentIdList = new ArrayList<>();
+            for(int i=0; i<studentIdArray.size(); i++){
 //            将学号格式化成Long,存入列表
-            String tmpStr = studentIdArray.getString(i);
-            studentIdList.add(Long.parseLong(tmpStr));
+                String tmpStr = studentIdArray.getString(i);
+                studentIdList.add(Long.parseLong(tmpStr));
+            }
+            responseObject = examinationService.deleteExamination(studentIdList);
+        }else{
+            responseObject.put("msg", "请至少输入一个学号!");
+            responseObject.put("code", 403);
         }
-        JSONObject responseObject = examinationService.deleteExamination(studentIdList);
         return responseObject;
     }
 }
